requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.9
django-backup==1.0.1
django-cors-headers==3.10.0
django-dbbackup==3.3.0
django-phonenumber-field==6.0.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
imageio==2.10.4
itypes==1.2.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
phonenumbers==8.12.37
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyWavelets==1.2.0
requests==2.26.0
requests-oauthlib==1.3.0
scikit-image==0.18.3
scipy==1.7.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tifffile==2021.11.2
uritemplate==4.1.1
urllib3==1.26.7
